+++ src/pages/_app.tsx
@@ -1,21 +1,7 @@
 import type { AppProps } from 'next/app';
+import { ChakraProvider } from '@chakra-ui/react';
 
-import { ChakraProvider, extendTheme } from '@chakra-ui/react';
-
-const theme = extendTheme({
-  colors: {
-    black: '#242529',
-    point: '#ffe867',
-    waggle: {
-      gray: {
-        600: '#dcdcdd',
-        700: '#5c6873',
-        800: '#4d4f57',
-        900: '#36373f',
-      },
-    },
-  },
-});
+import theme from '@/styles/theme';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
